feat(terminal): replace ghostty-web with xterm.js + WebGL renderer
Swap the WASM-based ghostty-web terminal renderer for xterm.js v6 with the
WebGL addon for GPU-accelerated rendering. DOM renderer is the automatic
fallback when WebGL2 is unavailable or the context is lost.

Changes:
- Remove ghostty-web dependency; add @xterm/xterm, @xterm/addon-webgl,
  @xterm/addon-fit, @xterm/addon-unicode11
- Delete GhosttyTerminal.tsx (WASM init singleton, ghostty Terminal class)
- Add XTermTerminal.tsx — structurally identical to GhosttyTerminal but
  synchronous (no WASM), with WebGL addon + onContextLoss fallback,
  Unicode 11 activation, and One Dark theme
- Update TerminalPane.tsx — import/type/JSX swap only, zero logic changes
- Import @xterm/xterm/css/xterm.css in main.tsx for correct rendering

The PTY data flow, resize handling, tab lifecycle, and handle shape
({ write, focus }) are all unchanged.
